+++ client/globals.ts
@@ -1,0 +1,4 @@
+import { HashRouter as i_0, BrowserRouter as i_1, Route as i_2, Routes as i_3, useMatch as i_4, useSearchParams as i_5, useLocation as i_6, useParams as i_7, Outlet as i_8 } from './routing';
+
+
+(window as any).PretRouter = {'HashRouter': i_0, 'BrowserRouter': i_1, 'Route': i_2, 'Routes': i_3, 'useMatch': i_4, 'useSearchParams': i_5, 'useLocation': i_6, 'useParams': i_7, 'Outlet': i_8};